requirements.txt
absl-py
numpy
pandas
tensorflow-datasets
jax
jaxlib
neural-tangents